+++ pkg/cluster/internal/create/actions/createworker/aws.go
@@ -254,6 +254,7 @@ func (b *AWSBuilder) installCSI(n nodes.Node, k string, privateParams PrivatePar
 	if getManifestErr != nil {
 		return errors.Wrap(getManifestErr, "failed to generate "+csiName+"-csi helm values")
 	}
+  
 	c := "echo '" + csiHelmValues + "' > " + csiValuesFile
 	_, err := commons.ExecuteCommand(n, c, 5, 3)
 	if err != nil {
@@ -330,6 +331,7 @@ spec:
 	// Create the eks.config file in the container
 	eksConfigPath := "/kind/eks.config"
 	c = "echo \"" + eksConfigData + "\" > " + eksConfigPath
+
 	_, err = commons.ExecuteCommand(n, c, 5, 3)
 	if err != nil {
 		return errors.Wrap(err, "failed to create eks.config")
@@ -337,6 +339,7 @@ spec:
 
 	// Run clusterawsadm with the eks.config file previously created (this will create or update the CloudFormation stack in AWS)
 	c = "clusterawsadm bootstrap iam create-cloudformation-stack --config " + eksConfigPath
+
 	_, err = commons.ExecuteCommand(n, c, 5, 3, envVars)
 	if err != nil {
 		return errors.Wrap(err, "failed to run clusterawsadm")
@@ -400,12 +403,14 @@ func (b *AWSBuilder) configureStorageClass(n nodes.Node, k string) error {
 	if b.capxManaged {
 		// Remove annotation from default storage class
 		c = "kubectl --kubeconfig " + k + ` get sc -o jsonpath='{.items[?(@.metadata.annotations.storageclass\.kubernetes\.io/is-default-class=="true")].metadata.name}'`
+
 		output, err := commons.ExecuteCommand(n, c, 5, 3)
 		if err != nil {
 			return errors.Wrap(err, "failed to get default storage class")
 		}
 		if strings.TrimSpace(output) != "" && strings.TrimSpace(output) != "No resources found" {
 			c = "kubectl --kubeconfig " + k + " annotate sc " + strings.TrimSpace(output) + " " + defaultScAnnotation + "-"
+
 			_, err = commons.ExecuteCommand(n, c, 5, 3)
 			if err != nil {
 				return errors.Wrap(err, "failed to remove annotation from default storage class")
@@ -467,6 +472,7 @@ func (b *AWSBuilder) postInstallPhase(n nodes.Node, k string) error {
 	var coreDNSPDBName = "coredns"
 
 	c := "kubectl --kubeconfig " + kubeconfigPath + " get pdb " + coreDNSPDBName + " -n kube-system"
+
 	_, err := commons.ExecuteCommand(n, c, 5, 3)
 	if err != nil {
 		err = installCorednsPdb(n)