+++ pkg/commons/cluster.go
@@ -143,7 +143,6 @@ type KeosSpec struct {
 
 type Keos struct {
 	Flavour string `yaml:"flavour,omitempty"`
-	Version string `yaml:"version,omitempty"`
 }
 
 type Networks struct {